Ensure view compact file removed on crash
Ugly hack stores view compaction file in the process dictionary of the
compacting process.

The other problem is that the assertion that it's a regular file before
the induced crash fails, presumably because the compaction has either
not started yet, or already finished, but it's clearly pretty racy.

In short, I'm not sure how much value there is in this approach.